client with token

+++ blueprints/song/resources.py
@@ -3,6 +3,8 @@ from flask import Blueprint
 from flask_restful import Api, Resource, reqparse, request, marshal
 from . import *
 import requests
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from ..client import Clients
 
 bp_song = Blueprint('song', __name__)
 api = Api(bp_song)
@@ -10,43 +12,59 @@ class SongResource(Resource):
     wio_host = 'http://ws.audioscrobbler.com/2.0/'
     wio_apikey = '27f9a43dd663f505b4e03d41db4d4d51'
 
+    @jwt_required
     def get(self, id=None):
-        if id is None:
-            parser = reqparse.RequestParser()
-            parser.add_argument('p', type=int, location='args', default=1)
-            parser.add_argument('rp', type=int, location='args', default=5)
-            parser.add_argument('genre', location='args')
-            parser.add_argument('artist', location='args')
-            args = parser.parse_args()
+        user = get_jwt_identity()
+        identity = marshal(user, Clients.response_field)
 
-            offset = (args['p'] * args['rp']) - args['rp']
+        if identity['status'] == 'user' or 'admin':
+            if id is None:
+                parser = reqparse.RequestParser()
+                parser.add_argument('p', type=int, location='args', default=1)
+                parser.add_argument('rp', type=int, location='args', default=5)
+                parser.add_argument('genre', location='args')
+                parser.add_argument('artist', location='args')
+                args = parser.parse_args()
 
-            qry = Songs.query
+                offset = (args['p'] * args['rp']) - args['rp']
 
-            if args['genre'] is not None:
-                qry = qry.filter_by(title=args['genre'])
+                qry = Songs.query
 
-            rows = []
-            for row in qry.limit(args['rp']).offset(offset).all():
-                rows.append(marshal(row, Songs.response_fields))
-            return marshal(rows, Songs.response_fields), 200, { 'Content-Type': 'application/json' }
-        
-        else:
-            qry = Songs.query.get(id)
-            if qry is not None:
-                return marshal(qry, Songs.response_fields), 200, { 'Content-Type': 'application/json' }
+                if args['genre'] is not None:
+                    qry = qry.filter_by(title=args['genre'])
+
+                rows = []
+                for row in qry.limit(args['rp']).offset(offset).all():
+                    rows.append(marshal(row, Songs.response_fields))
+                return marshal(rows, Songs.response_fields), 200, { 'Content-Type': 'application/json' }
+
             else:
-                return "Data Not Found", 200, { 'Content-Type': 'application/json' }
+                qry = Songs.query.get(id)
+                if qry is not None:
+                    return marshal(qry, Songs.response_fields), 200, { 'Content-Type': 'application/json' }
+                else:
+                    return "Data Not Found", 200, { 'Content-Type': 'application/json' }
+        else:
+            return 'UNAUTORIZED', 500, { 'Content-Type': 'application/json' }
 
+    @jwt_required
     def delete(self, id):
         qry = Songs.query.get(id)
-        if qry is not None:
-            db.session.delete(qry)
-            db.session.commit()
-            return "Data Deleted", 200, { 'Content-Type': 'application/json' }
-        else :
-            return "Data Not Found", 404, { 'Content-Type': 'application/json' }
 
+        user = get_jwt_identity()
+        identity = marshal(user, Clients.response_field)
+
+        if identity['status'] == 'admin':
+            if qry is not None:
+                db.session.delete(qry)
+                db.session.commit()
+                return "Data Deleted", 200, { 'Content-Type': 'application/json' }
+            else :
+                return "Data Not Found", 404, { 'Content-Type': 'application/json' }
+        else:
+            return 'UNAUTORIZED', 500, { 'Content-Type': 'application/json' }
+
+    @jwt_required
     def put(self, id):
         parser = reqparse.RequestParser()
         parser.add_argument('title', location='json', required=True)
@@ -55,15 +73,24 @@ class SongResource(Resource):
         args = parser.parse_args()
 
         qry = Songs.query.get(id)
-        if qry is not None:  
-            qry.title=args['title']
-            qry.artist=args['artist']
-            qry.genre=args['genre']    
-            db.session.commit()
-            return marshal(qry, Songs.response_fields), 200, { 'Content-Type': 'application/json' }
-        else :
-            return "Data Not Found", 404, { 'Content-Type': 'application/json' }
 
+        user = get_jwt_identity()
+        identity = marshal(user, Clients.response_field)
+
+        if identity['status'] == 'admin':
+
+            if qry is not None:  
+                qry.title=args['title']
+                qry.artist=args['artist']
+                qry.genre=args['genre']    
+                db.session.commit()
+                return marshal(qry, Songs.response_fields), 200, { 'Content-Type': 'application/json' }
+            else :
+                return "Data Not Found", 404, { 'Content-Type': 'application/json' }
+        else:
+            return 'UNAUTORIZED', 500, { 'Content-Type': 'application/json' }
+
+    @jwt_required
     def post(self):
         qry = Songs.query
         for data in qry:
@@ -76,17 +103,24 @@ class SongResource(Resource):
 
         args = parser.parse_args()
         list_genre = ["rock", "electronic", "seen live", "alternative", "indie", "pop", "female vocalists", "metal", "alternative rock", "jazz", "classic rock", "ambient", "experimental", "folk", "punk", "indie rock", "hard rock", "Hip-Hop", "instrumental", "singer-songwriter", "black metal", 	"dance", "80s", "Progressive rock", "death metal", "heavy metal", "hardcore", "british", "soul", "chillout", "electronica", "Classical", "industrial", "Soundtrack", "rap", "blues", "punk rock", "thrash metal", "90s", "acoustic", "metalcore", "psychedelic", "post-rock", "Progressive metal", "german", "funk", "hip hop", "new wave", "trance"]
-        for genre in list_genre :
-            rq = requests.get(self.wio_host, params={'method': args['method'], 'tag' : genre, 'format' : args['format'], 'api_key': self.wio_apikey})
-            song = rq.json()
-            for data in song['tracks']['track']:
-                songs = Songs(None, data['name'], data['artist']['name'], genre)
-                db.session.add(songs)
-                db.session.commit()
-        return marshal(songs, Songs.response_fields), 200, { 'Content-Type': 'application/json' }
+        
+        user = get_jwt_identity()
+        identity = marshal(user, Clients.response_field)
+
+        if identity['status'] == 'admin':
+            for genre in list_genre :
+                rq = requests.get(self.wio_host, params={'method': 'tag.gettoptracks', 'tag' : genre, 'format' : 'json', 'api_key': self.wio_apikey})
+                song = rq.json()
+                for data in song['tracks']['track']:
+                    songs = Songs(None, data['name'], data['artist']['name'], genre)
+                    db.session.add(songs)
+                    db.session.commit()
+            return 'SUCCESS', 200, { 'Content-Type': 'application/json' }
+        else:
+            return 'UNAUTORIZED', 500, { 'Content-Type': 'application/json' }
 
 class SongResourceAdmin(Resource):
-
+    @jwt_required
     def post(self):
         parser = reqparse.RequestParser()
         parser.add_argument('title', location='json', required=True)
@@ -94,12 +128,17 @@ class SongResourceAdmin(Resource):
         parser.add_argument('genre', location='json', required=True)
         args = parser.parse_args()
 
-        song = Songs(None, args['title'], args['artist'], args['genre'])
-        db.session.add(song)
-        db.session.commit()
+        user = get_jwt_identity()
+        identity = marshal(user, Clients.response_field)
 
-        return marshal(song, Songs.response_fields), 200, { 'Content-Type': 'application/json' }
+        if identity['status'] == 'admin':
+            song = Songs(None, args['title'], args['artist'], args['genre'])
+            db.session.add(song)
+            db.session.commit()
 
+            return marshal(song, Songs.response_fields), 200, { 'Content-Type': 'application/json' }
+        else:
+            return 'UNAUTORIZED', 500, { 'Content-Type': 'application/json' }
 
 api.add_resource(SongResource, '/song', '/song/<int:id>')
 api.add_resource(SongResourceAdmin, '/admin')
